add wrapper for setting page
