+++ .idea/workspace.xml
@@ -2,13 +2,11 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="23dd3188-ad99-4626-a6fc-8d6a150d9de3" name="Changes" comment="request user details">
-      <change afterPath="$PROJECT_DIR$/src/services/localStorage.ts" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/settings/WrapperForSettingPage.tsx" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/auth/components/logOutFor/LogOutButton.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/auth/components/logOutFor/LogOutButton.tsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/auth/components/loginForm/LoginForm.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/auth/components/loginForm/LoginForm.tsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/redux/authSlice.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/redux/authSlice.tsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/services/hondaApi.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/services/hondaApi.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/router/root.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/router/root.tsx" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/settings/SettingsPage.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/settings/SettingsPage.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/settings/constants.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/settings/constants.ts" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
