+++ src/settings/constants.ts
@@ -1,1 +1,2 @@
 export const placeWhereCar = 'Машина в гараже';
+export const carProvider = 'carProvider';
